+++ examples/campaign_management/get_all_disapproved_ads.pl
@@ -25,10 +25,11 @@ use lib "$Bin/../../lib";
 use Google::Ads::GoogleAds::Client;
 use Google::Ads::GoogleAds::Utils::GoogleAdsHelper;
 use Google::Ads::GoogleAds::Utils::SearchGoogleAdsIterator;
-use Google::Ads::GoogleAds::V17::Enums::PolicyApprovalStatusEnum
+use Google::Ads::GoogleAds::V18::Enums::PolicyApprovalStatusEnum
   qw(DISAPPROVED);
 use
-  Google::Ads::GoogleAds::V17::Services::GoogleAdsService::SearchGoogleAdsRequest;
+  Google::Ads::GoogleAds::V18::Services::GoogleAdsService::SearchGoogleAdsRequest;
+use Google::Ads::GoogleAds::V18::Services::GoogleAdsService::SearchSettings;
 
 use Getopt::Long qw(:config auto_help);
 use Pod::Usage;
@@ -58,12 +59,15 @@ sub get_all_disapproved_ads {
 
   # Create a search Google Ads request.
   my $search_request =
-    Google::Ads::GoogleAds::V17::Services::GoogleAdsService::SearchGoogleAdsRequest
+    Google::Ads::GoogleAds::V18::Services::GoogleAdsService::SearchGoogleAdsRequest
     ->new({
-      customerId              => $customer_id,
-      query                   => $search_query,
-      returnTotalResultsCount => "true"
-    });
+      customerId     => $customer_id,
+      query          => $search_query,
+      searchSettings =>
+        Google::Ads::GoogleAds::V18::Services::GoogleAdsService::SearchSettings
+        ->new({
+          returnTotalResultsCount => true
+        })});
 
   # Get the GoogleAdsService.
   my $google_ads_service = $api_client->GoogleAdsService();